fix(ap.js, state.js): handle RegEx errors and moves state to its own file
provide default values for each RegEx in case the given search and filters throw errors. We can also
use accessors on the state object to prevent accidentally trying to set values directly to a
computed property. Instead, use setters to make the access points for updating the state explicit.
This also allows us to compute properties as needed.