docs: ✏️ Remove warning blob about branch name change in README
The warning blob about a branch name change has been removed from the
README file. This warning was added when the repo still had a
development branch, and contributors were asked to send pull requests to
that branch. However, since the development branch has been dropped for
a while now, the warning is no longer relevant.

BREAKING CHANGE: 🧨 N/A

✅ Closes: 491